+++ Geometry3D.cpp
@@ -405,9 +405,9 @@ private:
   double invNorm;
 
   void normalize() {
-    this->invNorm = 1.0 / sqrt(this->A * this->A
-                             + this->B * this->B
-                             + this->C * this->C);
+    this->invNorm = 1.0 / sqrt((double)this->A * this->A
+                             + (double)this->B * this->B
+                             + (double)this->C * this->C);
   }
 
 public:
